split functions out of vira-dropdown file, organize dropdown files
