+++ packages/vira/src/elements/index.ts
@@ -1,10 +1,10 @@
 /** This file is automatically updated by update-index-exports.ts */
 
 export * from './define-vira-element';
+export * from './dropdown/vira-dropdown-item.element';
+export * from './dropdown/vira-dropdown.element';
 export * from './vira-button.element';
 export * from './vira-collapsible-wrapper.element';
-export * from './vira-dropdown-item.element';
-export * from './vira-dropdown.element';
 export * from './vira-icon.element';
 export * from './vira-image.element';
 export * from './vira-input.element';
